beamforming: pass fapi beam id to oru
change default beam id from -1 to 0 to be consistent with ORAN.
update phytest scheduler to send PDSCH on differnt beams beased on SSB index
beam allocation.

+++ openair2/LAYER2/NR_MAC_gNB/nr_mac_gNB.h
@@ -756,7 +756,7 @@ typedef struct {
   NR_UE_NR_Capability_t *capability;
   measgap_config_t measgap_config;
   // UE selected beam index
-  uint8_t UE_beam_index;
+  uint16_t UE_beam_index;
   float ul_thr_ue;
   float dl_thr_ue;
   long pdsch_HARQ_ACK_Codebook;
